[SCRIPT-5] Add fetching of existing github labels and optimize create and delete api calls (#13)
* feat: add fetching of existing github labels

* chore: lint

* feat: add logic to update label if label already exists

* chore: rename method _load_labels to more descriptive _load_labels_from_config

* chore: when logging loading labels from config error, label index starts at 1

* feat: add logging for fetching list of github labels

+++ scripts/setup_issue_label.py
@@ -65,7 +65,11 @@ class GithubIssueLabel:
             "Accept": "application/vnd.github+json",
             "X-GitHub-Api-Version": "2022-11-28",
         }
-        self._labels: list[dict[str, str]] = self._load_labels()
+        self._github_labels: list[dict[str, str]] = self._fetch_github_labels()
+        self._github_label_names: list[str] = [
+            github_label["name"] for github_label in self.github_labels
+        ]
+        self._labels: list[dict[str, str]] = self._load_labels_from_config()
 
     @property
     def url(self) -> str:
@@ -76,11 +80,41 @@ class GithubIssueLabel:
         return self._headers
 
     @property
+    def github_labels(self) -> list[dict[str, str]]:
+        return self._github_labels
+
+    @property
+    def github_label_names(self) -> list[str]:
+        return self._github_label_names
+
+    @property
     def labels(self) -> list[dict[str, str]]:
         return self._labels
 
-    def _load_labels(self) -> list[dict[str, str]]:
-        labels: list[dict[str, str]]
+    def _fetch_github_labels(self) -> list[dict[str, str]]:
+        res: Response = requests.get(
+            self.url,
+            headers=self.headers,
+        )
+
+        if res.status_code != 200:
+            logging.error(
+                f"Status {res.status_code}. Failed to fetch list of github labels"
+            )
+            sys.exit()
+
+        logging.info("Fetching list of github labels.")
+        return [
+            {
+                "name": github_label["name"],
+                "color": github_label["color"],
+                "description": github_label["description"],
+            }
+            for github_label in res.json()
+        ]
+
+    def _load_labels_from_config(self) -> list[dict[str, str]]:
+        labels: list[dict[str, str]] = []
         label_file: str = ""
 
         if os.path.isfile(LabelFile.YAML):
@@ -99,9 +133,35 @@ class GithubIssueLabel:
 
         with open(label_file, "r") as f:
             if label_file == LabelFile.YAML:
-                labels = yaml.safe_load(f)
+                for i, label in enumerate(yaml.safe_load(f), start=1):
+                    if not label.get("name"):
+                        logging.error(
+                            f"Name not found on `Label #{i}` with color `{label.get('color')}` and description `{label.get('description')}`."
+                        )
+                        sys.exit()
+
+                    labels.append(
+                        {
+                            "name": label.get("name"),
+                            "color": label.get("color", "").replace("#", ""),
+                            "description": label.get("description", ""),
+                        }
+                    )
 
         return labels
+
+    def _update_label(self, label: dict[str, str]) -> None:
+        url: str = f"{self.url}/{label['name']}"
+        label["new_name"] = label.pop("name")
+
+        res: Response = requests.patch(url, headers=self.headers, json=label)
+
+        if res.status_code != 200:
+            logging.error(
+                f"Status {res.status_code}. Failed to update label `{label['new_name']}`."
+            )
+        else:
+            logging.info(f"Label `{label['new_name']}` updated successfully.")
 
     def delete_default_labels(self) -> None:
         DEFAULT_LABEL_NAMES: list[str] = [
@@ -114,36 +174,45 @@ class GithubIssueLabel:
         ]
 
         for default_label_name in DEFAULT_LABEL_NAMES:
-            url: str = f"{self.url}/{default_label_name}"
+            if default_label_name in self.github_label_names:
+                url: str = f"{self.url}/{default_label_name}"
 
-            res: Response = requests.delete(
-                url,
-                headers=self.headers,
-            )
-
-            if res.status_code == 204:
-                logging.info(f"Label `{default_label_name}` deleted successfully.")
-            else:
-                logging.error(
-                    f"Status {res.status_code}. Failed to delete label `{default_label_name}`."
+                res: Response = requests.delete(
+                    url,
+                    headers=self.headers,
                 )
+
+                if res.status_code != 204:
+                    logging.error(
+                        f"Status {res.status_code}. Failed to delete label `{default_label_name}`."
+                    )
+                else:
+                    logging.info(f"Label `{default_label_name}` deleted successfully.")
 
     def create_labels(self) -> None:
         for label in self.labels:
-            label["color"] = label["color"].replace("#", "")
+            if label["name"] in self.github_label_names:
+                i: int = self.github_label_names.index(label["name"])
 
-            res: Response = requests.post(
-                self.url,
-                headers=self.headers,
-                json=label,
-            )
+                if (
+                    label["color"] != self.github_labels[i]["color"]
+                    or label["description"] != self.github_labels[i]["description"]
+                ):
+                    self._update_label(label)
 
-            if res.status_code == 201:
-                logging.info(f"Label `{label['name']}` created successfully.")
             else:
-                logging.error(
-                    f"Status {res.status_code}. Failed to create label `{label['name']}`."
+                res: Response = requests.post(
+                    self.url,
+                    headers=self.headers,
+                    json=label,
                 )
+
+                if res.status_code != 201:
+                    logging.error(
+                        f"Status {res.status_code}. Failed to create label `{label['name']}`."
+                    )
+                else:
+                    logging.info(f"Label `{label['name']}` created successfully.")
 
 
 def main() -> None:
